pyproject.toml
[project]
name = "d1"
version = "0.1.0"
description = "LangChain-powered desktop pet"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "PyQt6>=6.7",
    "langchain>=1.0",
    "langchain-community>=0.2",
    "langchain-openai>=0.1",
    "httpx>=0.26",
    "python-dotenv>=1.0",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = false
known-first-party = []
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
